chore(electrum): fromHeight/toHeight are integers

+++ bitcoin-regtest/bitcoin-regtest-core/src/main/java/org/tbk/bitcoin/regtest/BitcoindRegtestTestHelper.java
@@ -17,7 +17,7 @@ public final class BitcoindRegtestTestHelper {
     private static final Duration WAIT_FOR_SERVER_TIMEOUT = Duration.ofSeconds(30);
 
     public static void createLegacyDefaultWalletIfNecessary(BitcoinClient bitcoinJsonRpcClient) throws IOException {
-        boolean ready = bitcoinJsonRpcClient.waitForServer((int) WAIT_FOR_SERVER_TIMEOUT.toSeconds());
+        boolean ready = bitcoinJsonRpcClient.waitForServer(Math.toIntExact(WAIT_FOR_SERVER_TIMEOUT.toSeconds()));
         if (!ready) {
             throw new IllegalStateException("Server is not ready");
         }
@@ -36,7 +36,7 @@ public final class BitcoindRegtestTestHelper {
     }
 
     public static void createDescriptorWallet(BitcoinClient bitcoinJsonRpcClient, String walletName) throws IOException {
-        boolean ready = bitcoinJsonRpcClient.waitForServer((int) WAIT_FOR_SERVER_TIMEOUT.toSeconds());
+        boolean ready = bitcoinJsonRpcClient.waitForServer(Math.toIntExact(WAIT_FOR_SERVER_TIMEOUT.toSeconds()));
         if (!ready) {
             throw new IllegalStateException("Server is not ready");
         }
@@ -45,23 +45,23 @@ public final class BitcoindRegtestTestHelper {
     }
 
     private static synchronized void createWalletIfNecessary(BitcoinClient bitcoinJsonRpcClient, String walletName, boolean descriptors) throws IOException {
-            List<String> walletList = bitcoinJsonRpcClient.listWallets();
-            if (!walletList.contains(walletName)) {
-                // args for call to "createwallet". See https://developer.bitcoin.org/reference/rpc/createwallet.html
-                ImmutableMap<String, Optional<Object>> argsMap = ImmutableMap.<String, Optional<Object>>builder()
-                        .put("wallet_name", Optional.of(walletName))
-                        .put("disable_private_keys", Optional.of(false))
-                        .put("blank", Optional.of(false))
-                        .put("passphrase", Optional.empty())
-                        .put("avoid_reuse", Optional.of(false))
-                        .put("descriptors", Optional.of(descriptors))
-                        .put("load_on_startup", Optional.empty())
-                        .build();
+        List<String> walletList = bitcoinJsonRpcClient.listWallets();
+        if (!walletList.contains(walletName)) {
+            // args for call to "createwallet". See https://developer.bitcoin.org/reference/rpc/createwallet.html
+            ImmutableMap<String, Optional<Object>> argsMap = ImmutableMap.<String, Optional<Object>>builder()
+                    .put("wallet_name", Optional.of(walletName))
+                    .put("disable_private_keys", Optional.of(false))
+                    .put("blank", Optional.of(false))
+                    .put("passphrase", Optional.empty())
+                    .put("avoid_reuse", Optional.of(false))
+                    .put("descriptors", Optional.of(descriptors))
+                    .put("load_on_startup", Optional.empty())
+                    .build();
 
-                Object[] args = argsMap.values().stream().map(it -> it.orElse(null)).toArray();
-                Map<String, String> result = bitcoinJsonRpcClient.send("createwallet", args);
-                log.warn("Created default wallet: {}", result);
-            }
+            Object[] args = argsMap.values().stream().map(it -> it.orElse(null)).toArray();
+            Map<String, String> result = bitcoinJsonRpcClient.send("createwallet", args);
+            log.warn("Created default wallet: {}", result);
+        }
     }
 
     private BitcoindRegtestTestHelper() {
